adding recaptcha for register

+++ database/migrations/landlord/2022_05_25_073455_create_landlord_tenants_table.php
@@ -14,7 +14,7 @@ class CreateLandlordTenantsTable extends Migration
     public function up()
     {
         Schema::create('tenants', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->integer('id');
             $table->string('name');
             $table->string('domain')->unique();
             $table->string('database')->nullable();
